Updated attack packets and added !itemvac command.
git-svn-id: https://xp-dev.com/svn/MapleGame@53 22f4eea3-df18-4490-a9d7-79e6856d3c20

+++ ChannelServer/Maple/Skill.cs
@@ -319,7 +319,7 @@ namespace Loki.Maple
                     using (Packet outPacket = new Packet(MapleServerOperationCode.Cooldown))
                     {
                         outPacket.WriteInt(this.MapleID);
-                        outPacket.WriteShort((short)this.RemainingCooldownSeconds);
+                        outPacket.WriteInt((short)this.RemainingCooldownSeconds);
 
                         this.Character.Client.Send(outPacket);
                     }
@@ -329,7 +329,7 @@ namespace Loki.Maple
                         using (Packet outPacket = new Packet(MapleServerOperationCode.Cooldown))
                         {
                             outPacket.WriteInt(this.MapleID);
-                            outPacket.WriteShort(0);
+                            outPacket.WriteInt(0);
 
                             this.Character.Client.Send(outPacket);
                         }
